% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwc.R
\name{gwc}
\alias{gwc}
\title{Calculate generalised word count, for linear and quadratic effects}
\usage{
gwc(X.design, effects)
}
\arguments{
\item{effects}{Named list of numbers of polynomial effects: 'linear', 'quadratic'}

\item{\eqn{X.design}}{Design matrix}
}
\value{
Generalised word count -- a number
}
\description{
Given design matrix \eqn{X.design} and a named list of polynomial effects,
return the generalised word count.
Generalised case where quadratic effects of all factors can be included in the gwc.
}
