% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwc_linear.R
\name{gwc_linear}
\alias{gwc_linear}
\title{Calculate generalized word count for linear effects only, i.e. \eqn{b_k(k)}}
\usage{
gwc_linear(X.design, nlinear)
}
\arguments{
\item{nlinear}{Number of linear effects in the polynomial effect}

\item{\eqn{X.design}}{Design matrix}
}
\value{
Generalized word count
}
\description{
Given design matrix \eqn{X.design} and a number of linear effects,
return the generalized word count. Special case of gwc function.
}
