% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QB_I.R
\name{QB_I}
\alias{QB_I}
\title{Calculate Q_B criterion for the 1st order maximal model}
\usage{
QB_I(X.design, pi.linear)
}
\arguments{
\item{\eqn{X.design}}{Design matrix}

\item{\eqn{pi.linear}}{Prior probability of linear terms being in the best model}
}
\description{
Returns the \eqn{Q_B} value for the 1st order maximal model, as well as weights \eqn{\xi}
and generalized word count (\eqn{gwc})
}
