% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xi_1.R
\name{xi_1}
\alias{xi_1}
\title{Calculate \eqn{\xi} for the 1st order maximal model}
\usage{
xi_1(nfactors, pi)
}
\arguments{
\item{nfactors}{Number of factors, each is of two levels}

\item{pi}{Probability of each main effect being present in the best model}
}
\value{
A list of weights: \eqn{\xi.1} for main effects and {\xi.2} for pairs of main effects
}
\description{
Evaluate \eqn{\xi} - sum of probabilities of each model being the best, over the submodels of the
first-order maximal model containing (1) main effects of each factor (2) pairs of main effects
}
